+++ system/models/trds/stock.rb
@@ -1,0 +1,18 @@
+# encoding: utf-8
+module Trds
+  class Stock < Trst::Stock
+
+    has_many   :freights,     class_name: "Trds::FreightStock",         :inverse_of => :doc_stk, dependent: :destroy
+    belongs_to :unit,         class_name: "Trds::PartnerFirm::Unit",    :inverse_of => :stks
+
+    alias :file_name :name; alias :unit :unit_belongs_to
+
+    accepts_nested_attributes_for :freights,
+      reject_if: ->(attrs){ attrs[:qu].to_f == 0 && attrs[:pu].to_f == 0 },
+      allow_destroy: true
+
+    class << self
+    end # Class methods
+
+   end # Stock
+end #Trds