+++ CHANGELOG.md
@@ -1,3 +1,72 @@
+## v0.3.0
+
+* release date: **2014-09-16**
+   - bump version v0.3.0
+   - update CHANGELOG
+
+
+* **feature v0.2.992**, v0.3.0 preparation
+   - feature date: **2014-09-16**
+   - bump version v0.2.992
+   - update CHANGELOG
+   - i18n `partner_firms` update
+   - views `Trds::PartnerFirm` CRUD
+   - models generic skeletons, update
+   - assets js, update
+
+
+* **feature v0.2.991**, v0.3.0 preparation, close [issue #1][#1]
+   - feature date: **2014-08-20**
+   - bump version v0.2.991
+   - update CHANGELOG
+   - lib `version.rb` rename (typo)
+   - config `firm.yml` update
+   - assets css,  use unique style in modules [kfl62/trst_sys_sinatra#12][trst_main#12]
+   - views `layout`, use unique layout in `public` controller [kfl62/trst_sys_sinatra#15][trst_main#15]
+
+
+* **feature v0.1.3**, `Trds::PartnerFirm` CRUD
+   - feature date: **2014-04-02**
+   - bump version v0.1.3
+   - update CHANGELOG
+   - views `Trds::PartnerFirm` and related (address, person etc.) CRUD
+   - i18n `trds.yml` skeleton, `partner_firm.yml`
+   - assets js, `Trds.desk.partner_firm` skeleton
+   - assets js, `Trds` main module and helpers
+   - cleanup (.gitignore)
+
+
+* **feature v0.1.2**, mongoid models, base
+   - feature date: **2014-04-02**
+   - bump version v0.1.2
+   - update CHANGELOG
+   - add model `Trds::Cache`
+   - add model `Trds::Cassation`
+   - add model `Trds::Consumption`
+   - add model `Trds::DeliveryNote`
+   - add model `Trds::Freight`
+   - add model `Trds::FreightIn`
+   - add model `Trds::FreightOut`
+   - add model `Trds::FreightStock`
+   - add model `Trds::Grn`
+   - add model `Trds::Hmrs`
+   - add model `Trds::Invoice`
+   - add model `Trds::PartnerFirm`
+   - add model `Trds::Payment`
+   - add model `Trds::Stock`
+   - add model `Trds::User`
+   - cleanup (.gitignore)
+
+
+* **feature v0.1.1**, public/system pages, login etc.
+   - feature date: **2014-03-31**
+   - bump version v0.1.1
+   - update CHANGELOG
+   - config `firm.yml` minor changes
+   - views layout, public page
+   - assets css, cleanup, base files
+
+
 ## v0.1.0
 
 release date: **2014-03-31**
@@ -9,3 +78,48 @@ release date: **2014-03-31**
 * Add version `lib/trds/version.rb`
 
 * Add skeleton and common files, config, artwork, i18n etc.
+
+
+[trst_main#15]: https://github.com/kfl62/trst_sys_sinatra/issues/15
+[trst_main#12]: https://github.com/kfl62/trst_sys_sinatra/issues/12
+[#1]: https://github.com/kfl62/trst_sys_trds/issues/1
+[#2]: https://github.com/kfl62/trst_sys_trds/issues/2
+[#3]: https://github.com/kfl62/trst_sys_trds/issues/3
+[#4]: https://github.com/kfl62/trst_sys_trds/issues/4
+[#5]: https://github.com/kfl62/trst_sys_trds/issues/5
+[#6]: https://github.com/kfl62/trst_sys_trds/issues/6
+[#7]: https://github.com/kfl62/trst_sys_trds/issues/7
+[#8]: https://github.com/kfl62/trst_sys_trds/issues/8
+[#9]: https://github.com/kfl62/trst_sys_trds/issues/9
+[#10]: https://github.com/kfl62/trst_sys_trds/issues/10
+[#11]: https://github.com/kfl62/trst_sys_trds/issues/11
+[#12]: https://github.com/kfl62/trst_sys_trds/issues/12
+[#13]: https://github.com/kfl62/trst_sys_trds/issues/13
+[#14]: https://github.com/kfl62/trst_sys_trds/issues/14
+[#15]: https://github.com/kfl62/trst_sys_trds/issues/15
+[#16]: https://github.com/kfl62/trst_sys_trds/issues/16
+[#17]: https://github.com/kfl62/trst_sys_trds/issues/17
+[#18]: https://github.com/kfl62/trst_sys_trds/issues/18
+[#19]: https://github.com/kfl62/trst_sys_trds/issues/19
+[#20]: https://github.com/kfl62/trst_sys_trds/issues/20
+[#21]: https://github.com/kfl62/trst_sys_trds/issues/21
+[#22]: https://github.com/kfl62/trst_sys_trds/issues/22
+[#23]: https://github.com/kfl62/trst_sys_trds/issues/23
+[#24]: https://github.com/kfl62/trst_sys_trds/issues/24
+[#25]: https://github.com/kfl62/trst_sys_trds/issues/25
+[#26]: https://github.com/kfl62/trst_sys_trds/issues/26
+[#27]: https://github.com/kfl62/trst_sys_trds/issues/27
+[#28]: https://github.com/kfl62/trst_sys_trds/issues/28
+[#29]: https://github.com/kfl62/trst_sys_trds/issues/29
+[#30]: https://github.com/kfl62/trst_sys_trds/issues/30
+[#31]: https://github.com/kfl62/trst_sys_trds/issues/31
+[#32]: https://github.com/kfl62/trst_sys_trds/issues/32
+[#33]: https://github.com/kfl62/trst_sys_trds/issues/33
+[#34]: https://github.com/kfl62/trst_sys_trds/issues/34
+[#35]: https://github.com/kfl62/trst_sys_trds/issues/35
+[#36]: https://github.com/kfl62/trst_sys_trds/issues/36
+[#37]: https://github.com/kfl62/trst_sys_trds/issues/37
+[#38]: https://github.com/kfl62/trst_sys_trds/issues/38
+[#39]: https://github.com/kfl62/trst_sys_trds/issues/39
+[#40]: https://github.com/kfl62/trst_sys_trds/issues/40
+